+++ src/commands/index.ts
@@ -1,4 +1,5 @@
 export * from './installUpdate';
 export * from './pickVersion';
 export * from './pinVersion';
+export * from './removePackage';
 export * from './switchDepType';